// src/blocks/Form/types.ts
export interface Email {
  emailTo: string
  emailFrom: string
  cc?: string
  bcc?: string
  replyTo?: string
  subject: string
  message?: any // TODO: configure rich text type
}

export interface FormattedEmail {
  to: string
  cc?: string
  bcc?: string
  from: string
  subject: string
  html: string
  replyTo: string
  attachments?: {
    filename: string
    path: string
  }[]
}